トーストのUI調整
